+++ app/src/main/java/jp/ryuk/deglog/data/Diary.kt
@@ -9,7 +9,7 @@ data class Diary(
     var id: Long = 0L,
     var date: Long = System.currentTimeMillis(),
     var name: String = "",
-    var weight: Int? = null,
-    var length: Int? = null,
+    var weight: Float? = null,
+    var length: Float? = null,
     var memo: String? = null
 )
